qcacmn: Update TDLS opclass correctly to FW
Due to change in Opclass calculation in the new
regulatory component invalid opclass is returned for the
TDLS component. Update arguments to calculate opclass correctly
to regulatory component.

CRs-Fixed: 2325834

+++ umac/tdls/dispatcher/inc/wlan_tdls_public_structs.h
@@ -1140,7 +1140,7 @@ struct tdls_set_offchanmode {
  */
 struct tdls_set_secoffchanneloffset {
 	struct wlan_objmgr_vdev *vdev;
-	uint8_t offchan_offset;
+	int offchan_offset;
 	tdls_offchan_parms_callback callback;
 };
